Implemented authorization using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -63,6 +63,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)){
+            abort(403, 'Access Denied');
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -88,6 +91,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)){
+            abort(403, 'Access Denied');
+        }
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted!');
